requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djcli @ file:///Users/vince/Documents/PYTHON/djcli
drf-yasg==1.21.8
filelock==3.16.1
firebase-admin==6.6.0
fonttools==4.55.3
forex-python==1.8
frozendict==2.4.6
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
joblib==1.4.2
kombu==5.4.2
lxml==5.3.0
msgpack==1.1.0
multitasking==0.0.11
mysqlclient==2.2.7
numpy==2.2.1
openai==1.59.6
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydyf==0.11.0
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.1
PyPDF2==3.0.1
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stripe==11.4.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.9.0
yfinance==0.2.51
zopfli==0.2.3.post1
